Do not trap exits in the fast_resume code.
If we terminate with shutdown or normal, we believe we are in a stable
state, so in that case, lets write the fast_resume table to disk. Any
other state is outright friggin' dangerous to terminate from, so we
chose to play it safe and take the last good state on the disk that we
know of instead.